requirements.txt
click==8.0.1
ruamel.yaml==0.17.16
